DEV🛠️: Updated workflows and settings for Azure deployment
In this commit, several changes were made to improve the deployment process of the .NET Core application to Azure. The `FoodService.yml` file was removed, and the `azure-foodservice-mvc.yml` file was updated with changes to the workflow trigger, addition of GitHub Packages NuGet source, a new `Restore` step, and updated permissions for the `deploy` job. The `FoodService.sln` file was updated to reflect these changes. Lastly, the `BASE_URL_API` environment variable in `launchSettings.json` was updated for both `FoodService` and `FoodService.Homolog` profiles.